// lib/compass-api-client/src/generated/api.schemas.ts
export interface PlanFeatures {
  whiteLabel: boolean;
  multiAccreditorExport: boolean;
  customDomain: boolean;
}